configure: use 'command -v' instead of 'which'
'which' is an external command, that may or may not be installed in a
build environment. On the other hand, 'command -v' is a shell built-in
specified by POSIX, so it will always be available.

I hit this issue in a build environment where I knew 'meson' was in
$PATH, but still the configure script claimed it couldn't be found.
This fixes it.